FlipkartChatbot/requirements.txt
streamlit==1.33.0
groq>=0.9.0
python-dotenv==1.0.1
pyttsx3==2.90
SpeechRecognition==3.10.0
streamlit-mic-recorder==0.0.8
plotly==5.17.0
pandas>=2.2.0